refactor: Spoke 테스트 분리
